+++ coffee-web/src/main/java/me/dkist/iot/App.java
@@ -4,9 +4,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class CoffeeWebApplication {
+public class App {
 
 	public static void main(String[] args) {
-		SpringApplication.run(CoffeeWebApplication.class, args);
+		SpringApplication.run(App.class, args);
 	}
 }
