+++ coffee-web/src/main/java/me/dkist/iot/config/MongoDbConfig.java
@@ -1,4 +1,4 @@
-package me.dkist.iot.web;
+package me.dkist.iot.config;
 
 import java.net.UnknownHostException;
